Renderer Framework for Forms (new) - Duplicate script names fixed
Scripts with the same name, on different projects, were being overwritten. The consequence is that only one of the scripts with the same name will exist in the repository, while the others would be lost (as if deleted).

Solution: All scripts from the script project formscomposerdesigner, now have "Composer" as a suffix.

The table data for the freeStepWizard is added to the renderer customization.